Fixed import and made constructors package private
