+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/management/targettable/TargetTable.java
@@ -53,11 +53,11 @@ import org.eclipse.hawkbit.ui.management.event.TargetTableEvent;
 import org.eclipse.hawkbit.ui.management.event.TargetTableEvent.TargetComponentEvent;
 import org.eclipse.hawkbit.ui.management.state.ManagementUIState;
 import org.eclipse.hawkbit.ui.management.state.TargetTableFilters;
-import org.eclipse.hawkbit.ui.push.events.CancelTargetAssignmentEventContainer;
-import org.eclipse.hawkbit.ui.push.events.TargetCreatedEventContainer;
-import org.eclipse.hawkbit.ui.push.events.TargetDeletedEventContainer;
-import org.eclipse.hawkbit.ui.push.events.TargetInfoUpdateEventContainer;
-import org.eclipse.hawkbit.ui.push.events.TargetUpdatedEventContainer;
+import org.eclipse.hawkbit.ui.push.CancelTargetAssignmentEventContainer;
+import org.eclipse.hawkbit.ui.push.TargetCreatedEventContainer;
+import org.eclipse.hawkbit.ui.push.TargetDeletedEventContainer;
+import org.eclipse.hawkbit.ui.push.TargetInfoUpdateEventContainer;
+import org.eclipse.hawkbit.ui.push.TargetUpdatedEventContainer;
 import org.eclipse.hawkbit.ui.utils.AssignInstalledDSTooltipGenerator;
 import org.eclipse.hawkbit.ui.utils.HawkbitCommonUtil;
 import org.eclipse.hawkbit.ui.utils.SPDateTimeUtil;
