+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/push/DistributionSetTagDeletedEventContainer.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.hawkbit.ui.push.events;
+package org.eclipse.hawkbit.ui.push;
 
 import java.util.List;
 
@@ -19,7 +19,7 @@ import org.eclipse.hawkbit.repository.eventbus.event.DistributionSetTagDeletedEv
 public class DistributionSetTagDeletedEventContainer implements EventContainer<DistributionSetTagDeletedEvent> {
     private final List<DistributionSetTagDeletedEvent> events;
 
-    public DistributionSetTagDeletedEventContainer(final List<DistributionSetTagDeletedEvent> events) {
+    DistributionSetTagDeletedEventContainer(final List<DistributionSetTagDeletedEvent> events) {
         this.events = events;
     }
