+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/push/DistributionSetTagUpdatedEventContainer.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.hawkbit.ui.push.events;
+package org.eclipse.hawkbit.ui.push;
 
 import java.util.List;
 
@@ -19,7 +19,7 @@ import org.eclipse.hawkbit.repository.eventbus.event.DistributionSetTagUpdateEve
 public class DistributionSetTagUpdatedEventContainer implements EventContainer<DistributionSetTagUpdateEvent> {
     private final List<DistributionSetTagUpdateEvent> events;
 
-    public DistributionSetTagUpdatedEventContainer(final List<DistributionSetTagUpdateEvent> events) {
+    DistributionSetTagUpdatedEventContainer(final List<DistributionSetTagUpdateEvent> events) {
         this.events = events;
     }
