+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/push/RolloutGroupChangeEventContainer.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.hawkbit.ui.push.events;
+package org.eclipse.hawkbit.ui.push;
 
 import java.util.List;
 
@@ -19,7 +19,7 @@ import org.eclipse.hawkbit.repository.eventbus.event.RolloutGroupChangeEvent;
 public class RolloutGroupChangeEventContainer implements EventContainer<RolloutGroupChangeEvent> {
     private final List<RolloutGroupChangeEvent> events;
 
-    public RolloutGroupChangeEventContainer(final List<RolloutGroupChangeEvent> events) {
+    RolloutGroupChangeEventContainer(final List<RolloutGroupChangeEvent> events) {
         this.events = events;
     }
