+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/push/TargetInfoUpdateEventContainer.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.hawkbit.ui.push.events;
+package org.eclipse.hawkbit.ui.push;
 
 import java.util.List;
 
@@ -19,7 +19,7 @@ import org.eclipse.hawkbit.repository.eventbus.event.TargetInfoUpdateEvent;
 public class TargetInfoUpdateEventContainer implements EventContainer<TargetInfoUpdateEvent> {
     private final List<TargetInfoUpdateEvent> events;
 
-    public TargetInfoUpdateEventContainer(final List<TargetInfoUpdateEvent> events) {
+    TargetInfoUpdateEventContainer(final List<TargetInfoUpdateEvent> events) {
         this.events = events;
     }
