+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/push/TargetTagAssigmentResultEventContainer.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.hawkbit.ui.push.events;
+package org.eclipse.hawkbit.ui.push;
 
 import java.util.List;
 
@@ -19,7 +19,7 @@ import org.eclipse.hawkbit.repository.eventbus.event.TargetTagAssigmentResultEve
 public class TargetTagAssigmentResultEventContainer implements EventContainer<TargetTagAssigmentResultEvent> {
     private final List<TargetTagAssigmentResultEvent> events;
 
-    public TargetTagAssigmentResultEventContainer(final List<TargetTagAssigmentResultEvent> events) {
+    TargetTagAssigmentResultEventContainer(final List<TargetTagAssigmentResultEvent> events) {
         this.events = events;
     }
